pyproject.toml
[project]
name = "dialog_graph"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Matt Zhang", email = "set.stun@gmail.com"},
]
dependencies = [
    "py-gpt-interface>=0.8.2",
    "jinja2>=3.1.4",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "-e file:///${PROJECT_ROOT}/#egg=dialog-graph",
]
